Divided shim and useful functions
